Upd: src files organized and several slight fixes
