+++ src/Footer/index.tsx
@@ -1,0 +1,4 @@
+export * from './Footer';
+export * from './Filters';
+export * from './ClearCompleted';
+export * from './TodoCounter';
