+++ src/Header/index.tsx
@@ -1,0 +1,4 @@
+export * from './Header';
+export * from './HeaderInput';
+export * from './HeaderTitle';
+export * from './ToggleAllTodos';
