+++ src/Main/TodoList.tsx
@@ -4,7 +4,7 @@ type Props = {
   children: JSX.Element;
 };
 
-export const Main: React.FC<Props> = ({ children }) => (
+export const TodoList: React.FC<Props> = ({ children }) => (
   <section className="todoapp__main">
     {children}
   </section>
