+++ src/Main/index.tsx
@@ -1,0 +1,2 @@
+export * from './TodoList';
+export * from './TodoCard';
