+++ src/styles/todo.scss
@@ -75,12 +75,10 @@
   }
 
   &__edit {
+    position: relative;
     box-sizing: border-box;
-    width: 550px;
-    // line-height: 1.4em;
-    position: absolute;
-    left: 0;
-    padding: 11px 0 11px 55px;
+    width: 100%;
+    padding: 11px 0 11px 14px;
 
     font-size: inherit;
     line-height: inherit;
